+++ lib/metrics/namespaced.go
@@ -22,6 +22,12 @@ func Namespaced(t Type, ns string) Type {
 	}
 }
 
+// Unwrap to the underlying metrics type.
+// TODO: V4 make this standard for Type
+func (d namespacedWrapper) Unwrap() Type {
+	return d.t
+}
+
 //------------------------------------------------------------------------------
 
 func (d namespacedWrapper) GetCounter(path string) StatCounter {