+++ lib/processor/metric_test.go
@@ -1,12 +1,13 @@
 package processor
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/message"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 //------------------------------------------------------------------------------
@@ -47,16 +48,14 @@ func TestMetricBad(t *testing.T) {
 	conf := NewConfig()
 	conf.Type = "metric"
 	conf.Metric.Type = "bad type"
-	conf.Metric.Path = "some.path"
-	if _, err := New(conf, nil, log.Noop(), metrics.Noop()); err == nil {
-		t.Error("Expected error from bad type")
-	}
+	conf.Metric.Name = "some.path"
+	_, err := New(conf, nil, log.Noop(), metrics.Noop())
+	require.Error(t, err)
 
 	conf = NewConfig()
 	conf.Type = "metric"
-	if _, err := New(conf, nil, log.Noop(), metrics.Noop()); err == nil {
-		t.Error("Expected error from empty path")
-	}
+	_, err = New(conf, nil, log.Noop(), metrics.Noop())
+	require.Error(t, err)
 }
 
 func TestMetricCounter(t *testing.T) {
@@ -67,67 +66,11 @@ func TestMetricCounter(t *testing.T) {
 	conf := NewConfig()
 	conf.Type = "metric"
 	conf.Metric.Type = "counter"
-	conf.Metric.Path = "foo.bar"
+	conf.Metric.Name = "foo.bar"
 	conf.Metric.Value = "${!json(\"foo.bar\")}"
 
 	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	inputs := [][][]byte{
-		{},
-		{},
-		{
-			[]byte(`{}`),
-			[]byte(`{}`),
-		},
-		{
-			[]byte(`not even json`),
-		},
-		{},
-		{
-			[]byte(`{"foo":{"bar":"hello world"}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":{"baz":"hello world"}}}`),
-		},
-	}
-
-	expMetrics := map[string]int64{
-		"foo.bar": 7,
-	}
-
-	for _, i := range inputs {
-		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
-	}
-
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
-}
-
-func TestMetricCounterParts(t *testing.T) {
-	mockStats := &mockMetric{
-		values: map[string]int64{},
-	}
-
-	conf := NewConfig()
-	conf.Type = "metric"
-	conf.Metric.Type = "counter_parts"
-	conf.Metric.Path = "foo.bar"
-	conf.Metric.Value = "${!json(\"foo.bar\")}"
-
-	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, err)
 
 	inputs := [][][]byte{
 		{},
@@ -154,17 +97,11 @@ func TestMetricCounterParts(t *testing.T) {
 
 	for _, i := range inputs {
 		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
+		assert.Len(t, msg, 1)
+		assert.Nil(t, res)
 	}
 
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
+	assert.Equal(t, expMetrics, mockStats.values)
 }
 
 func TestMetricCounterBy(t *testing.T) {
@@ -175,13 +112,11 @@ func TestMetricCounterBy(t *testing.T) {
 	conf := NewConfig()
 	conf.Type = "metric"
 	conf.Metric.Type = "counter_by"
-	conf.Metric.Path = "foo.bar"
+	conf.Metric.Name = "foo.bar"
 	conf.Metric.Value = "${!json(\"foo.bar\")}"
 
 	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, err)
 
 	inputs := [][][]byte{
 		{},
@@ -189,6 +124,7 @@ func TestMetricCounterBy(t *testing.T) {
 		{
 			[]byte(`{"foo":{"bar":2}}`),
 			[]byte(`{}`),
+			[]byte(`{"foo":{"bar":3}}`),
 		},
 		{
 			[]byte(`not even json`),
@@ -205,22 +141,16 @@ func TestMetricCounterBy(t *testing.T) {
 	}
 
 	expMetrics := map[string]int64{
-		"foo.bar": 5,
+		"foo.bar": 8,
 	}
 
 	for _, i := range inputs {
 		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
+		assert.Len(t, msg, 1)
+		assert.Nil(t, res)
 	}
 
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
+	assert.Equal(t, expMetrics, mockStats.values)
 }
 
 func TestMetricGauge(t *testing.T) {
@@ -231,13 +161,11 @@ func TestMetricGauge(t *testing.T) {
 	conf := NewConfig()
 	conf.Type = "metric"
 	conf.Metric.Type = "gauge"
-	conf.Metric.Path = "foo.bar"
+	conf.Metric.Name = "foo.bar"
 	conf.Metric.Value = "${!json(\"foo.bar\")}"
 
 	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, err)
 
 	inputs := [][][]byte{
 		{},
@@ -251,120 +179,8 @@ func TestMetricGauge(t *testing.T) {
 		},
 		{
 			[]byte(`{"foo":{"bar":-5}}`),
+			[]byte(`{"foo":{"bar":7}}`),
 		},
-		{
-			[]byte(`{"foo":{"bar":"hello world"}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":{"baz":"hello world"}}}`),
-		},
-	}
-
-	expMetrics := map[string]int64{
-		"foo.bar": 5,
-	}
-
-	for _, i := range inputs {
-		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
-	}
-
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
-}
-
-func TestMetricTiming(t *testing.T) {
-	mockStats := &mockMetric{
-		values: map[string]int64{},
-	}
-
-	conf := NewConfig()
-	conf.Type = "metric"
-	conf.Metric.Type = "timing"
-	conf.Metric.Path = "foo.bar"
-	conf.Metric.Value = "${!json(\"foo.bar\")}"
-
-	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	inputs := [][][]byte{
-		{},
-		{},
-		{
-			[]byte(`{"foo":{"bar":5}}`),
-			[]byte(`{}`),
-		},
-		{
-			[]byte(`not even json`),
-		},
-		{
-			[]byte(`{"foo":{"bar":-5}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":"hello world"}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":{"baz":"hello world"}}}`),
-		},
-	}
-
-	expMetrics := map[string]int64{
-		"foo.bar": 5,
-	}
-
-	for _, i := range inputs {
-		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
-	}
-
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
-}
-
-func TestMetricCounterLabels(t *testing.T) {
-	mockStats := &mockMetric{
-		values: map[string]int64{},
-	}
-
-	conf := NewConfig()
-	conf.Type = "metric"
-	conf.Metric.Type = "counter"
-	conf.Metric.Path = "foo.bar"
-	conf.Metric.Value = "${!json(\"foo.bar\")}"
-	conf.Metric.Labels = map[string]string{
-		"batch_size": "${!batch_size()}",
-	}
-
-	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	inputs := [][][]byte{
-		{},
-		{},
-		{
-			[]byte(`{}`),
-			[]byte(`{}`),
-		},
-		{
-			[]byte(`not even json`),
-		},
-		{},
 		{
 			[]byte(`{"foo":{"bar":"hello world"}}`),
 		},
@@ -379,195 +195,14 @@ func TestMetricCounterLabels(t *testing.T) {
 
 	for _, i := range inputs {
 		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
+		assert.Len(t, msg, 1)
+		assert.Nil(t, res)
 	}
 
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
+	assert.Equal(t, expMetrics, mockStats.values)
 }
 
-func TestMetricCounterPartsLabels(t *testing.T) {
-	mockStats := &mockMetric{
-		values: map[string]int64{},
-	}
-
-	conf := NewConfig()
-	conf.Type = "metric"
-	conf.Metric.Type = "counter_parts"
-	conf.Metric.Path = "foo.bar"
-	conf.Metric.Value = "${!json(\"foo.bar\")}"
-	conf.Metric.Labels = map[string]string{
-		"batch_size": "${!batch_size()}",
-	}
-
-	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	inputs := [][][]byte{
-		{},
-		{},
-		{
-			[]byte(`{}`),
-			[]byte(`{}`),
-		},
-		{
-			[]byte(`not even json`),
-		},
-		{},
-		{
-			[]byte(`{"foo":{"bar":"hello world"}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":{"baz":"hello world"}}}`),
-		},
-	}
-
-	expMetrics := map[string]int64{
-		"foo.bar": 5,
-	}
-
-	for _, i := range inputs {
-		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
-	}
-
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
-}
-
-func TestMetricCounterByLabels(t *testing.T) {
-	mockStats := &mockMetric{
-		values: map[string]int64{},
-	}
-
-	conf := NewConfig()
-	conf.Type = "metric"
-	conf.Metric.Type = "counter_by"
-	conf.Metric.Path = "foo.bar"
-	conf.Metric.Value = "${!json(\"foo.bar\")}"
-	conf.Metric.Labels = map[string]string{
-		"batch_size": "${!batch_size()}",
-	}
-
-	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	inputs := [][][]byte{
-		{},
-		{},
-		{
-			[]byte(`{"foo":{"bar":2}}`),
-			[]byte(`{}`),
-		},
-		{
-			[]byte(`not even json`),
-		},
-		{
-			[]byte(`{"foo":{"bar":-2}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":3}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":{"baz":"hello world"}}}`),
-		},
-	}
-
-	expMetrics := map[string]int64{
-		"foo.bar": 5,
-	}
-
-	for _, i := range inputs {
-		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
-	}
-
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
-}
-
-func TestMetricGaugeLabels(t *testing.T) {
-	mockStats := &mockMetric{
-		values: map[string]int64{},
-	}
-
-	conf := NewConfig()
-	conf.Type = "metric"
-	conf.Metric.Type = "gauge"
-	conf.Metric.Path = "foo.bar"
-	conf.Metric.Value = "${!json(\"foo.bar\")}"
-	conf.Metric.Labels = map[string]string{
-		"batch_size": "${!batch_size()}",
-	}
-
-	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	inputs := [][][]byte{
-		{},
-		{},
-		{
-			[]byte(`{"foo":{"bar":5}}`),
-			[]byte(`{}`),
-		},
-		{
-			[]byte(`not even json`),
-		},
-		{
-			[]byte(`{"foo":{"bar":-5}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":"hello world"}}`),
-		},
-		{
-			[]byte(`{"foo":{"bar":{"baz":"hello world"}}}`),
-		},
-	}
-
-	expMetrics := map[string]int64{
-		"foo.bar": 5,
-	}
-
-	for _, i := range inputs {
-		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
-	}
-
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
-}
-
-func TestMetricTimingLabels(t *testing.T) {
+func TestMetricTiming(t *testing.T) {
 	mockStats := &mockMetric{
 		values: map[string]int64{},
 	}
@@ -575,16 +210,11 @@ func TestMetricTimingLabels(t *testing.T) {
 	conf := NewConfig()
 	conf.Type = "metric"
 	conf.Metric.Type = "timing"
-	conf.Metric.Path = "foo.bar"
+	conf.Metric.Name = "foo.bar"
 	conf.Metric.Value = "${!json(\"foo.bar\")}"
-	conf.Metric.Labels = map[string]string{
-		"batch_size": "${!batch_size()}",
-	}
 
 	proc, err := New(conf, nil, log.Noop(), metrics.WrapFlat(mockStats))
-	if err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, err)
 
 	inputs := [][][]byte{
 		{},
@@ -598,6 +228,7 @@ func TestMetricTimingLabels(t *testing.T) {
 		},
 		{
 			[]byte(`{"foo":{"bar":-5}}`),
+			[]byte(`{"foo":{"bar":7}}`),
 		},
 		{
 			[]byte(`{"foo":{"bar":"hello world"}}`),
@@ -608,22 +239,14 @@ func TestMetricTimingLabels(t *testing.T) {
 	}
 
 	expMetrics := map[string]int64{
-		"foo.bar": 5,
+		"foo.bar": 7,
 	}
 
 	for _, i := range inputs {
 		msg, res := proc.ProcessMessage(message.New(i))
-		if exp, act := 1, len(msg); exp != act {
-			t.Errorf("Wrong count of resulting messages: %v != %v", act, exp)
-		}
-		if res != nil {
-			t.Error(res.Error())
-		}
+		assert.Len(t, msg, 1)
+		assert.Nil(t, res)
 	}
 
-	if !reflect.DeepEqual(expMetrics, mockStats.values) {
-		t.Errorf("Wrong result: %v != %v", mockStats.values, expMetrics)
-	}
+	assert.Equal(t, expMetrics, mockStats.values)
 }
-
-//------------------------------------------------------------------------------