+++ lib/processor/util.go
@@ -154,6 +154,29 @@ func ClearFail(part types.Part) {
 
 //------------------------------------------------------------------------------
 
+func iterateParts(
+	parts []int, msg types.Message,
+	iter func(int, types.Part) error,
+) error {
+	exec := func(i int) error {
+		return iter(i, msg.Get(i))
+	}
+	if len(parts) == 0 {
+		for i := 0; i < msg.Len(); i++ {
+			if err := exec(i); err != nil {
+				return err
+			}
+		}
+	} else {
+		for _, i := range parts {
+			if err := exec(i); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
 // IteratePartsWithSpan iterates the parts of a message according to a slice of
 // indexes (if empty all parts are iterated) and calls a func for each part
 // along with a tracing span for that part. If an error is returned the part is